Moved classes to correct folder/namespace.
