+++ NitroxClient/Communication/Packets/Processors/ConstructorBeginCraftingProcessor.cs
@@ -1,13 +1,13 @@
-﻿using NitroxClient.Communication.Packets.Processors.Abstract;
-using NitroxModel.DataStructures;
-using NitroxModel.DataStructures.Util;
-using NitroxModel.Helper;
+﻿using System.Collections.Generic;
+using System.Reflection;
+using NitroxClient.Communication.Packets.Processors.Abstract;
 using NitroxClient.GameLogic.Helper;
 using NitroxClient.Unity.Helper;
+using NitroxModel.DataStructures.GameLogic;
+using NitroxModel.DataStructures.Util;
+using NitroxModel.Helper;
 using NitroxModel.Logger;
 using NitroxModel.Packets;
-using System.Collections.Generic;
-using System.Reflection;
 using UnityEngine;
 using static NitroxClient.GameLogic.Helper.TransientLocalObjectManager;
 
@@ -21,14 +21,14 @@ namespace NitroxClient.Communication.Packets.Processors
         {
             GameObject gameObject = GuidHelper.RequireObjectFrom(packet.ConstructorGuid);
             Crafter crafter = gameObject.RequireComponentInChildren<Crafter>(true);
-                                    
+
             MethodInfo onCraftingBegin = typeof(Crafter).GetMethod("OnCraftingBegin", BindingFlags.NonPublic | BindingFlags.Instance);
             Validate.NotNull(onCraftingBegin);
             onCraftingBegin.Invoke(crafter, new object[] { packet.TechType, packet.Duration }); //TODO: take into account latency for duration   
 
             Optional<object> opConstructedObject = TransientLocalObjectManager.Get(TransientObjectType.CONSTRUCTOR_INPUT_CRAFTED_GAMEOBJECT);
 
-            if(opConstructedObject.IsPresent())
+            if (opConstructedObject.IsPresent())
             {
                 GameObject constructedObject = (GameObject)opConstructedObject.Get();
                 GuidHelper.SetNewGuid(constructedObject, packet.ConstructedItemGuid);
@@ -43,11 +43,11 @@ namespace NitroxClient.Communication.Packets.Processors
 
         private void SetInteractiveChildrenGuids(GameObject constructedObject, List<InteractiveChildObjectIdentifier> interactiveChildIdentifiers)
         {
-            foreach(InteractiveChildObjectIdentifier childIdentifier in interactiveChildIdentifiers)
+            foreach (InteractiveChildObjectIdentifier childIdentifier in interactiveChildIdentifiers)
             {
-                UnityEngine.Transform transform = constructedObject.transform.Find(childIdentifier.GameObjectNamePath);
+                Transform transform = constructedObject.transform.Find(childIdentifier.GameObjectNamePath);
 
-                if(transform != null)
+                if (transform != null)
                 {
                     GameObject gameObject = transform.gameObject;
                     GuidHelper.SetNewGuid(gameObject, childIdentifier.Guid);
