+++ NitroxClient/GameLogic/Constructor.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using NitroxClient.Communication;
 using NitroxClient.GameLogic.Helper;
-using NitroxModel.DataStructures;
+using NitroxModel.DataStructures.GameLogic;
 using NitroxModel.DataStructures.Util;
 using NitroxModel.Logger;
 using NitroxModel.Packets;
