+++ NitroxModel/Extensions.cs
@@ -55,6 +55,7 @@ namespace NitroxModel
                 sb.Append(prefix);
                 sb.Append(i);
             }
+
             return sb.ToString();
         }
     }
