+++ NitroxServer/GameLogic/EntityManager.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using NitroxModel.DataStructures;
+using NitroxModel.DataStructures.GameLogic;
 using NitroxModel.GameLogic;
 using NitroxModel.Helper;
 using UnityEngine;
